+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+- Added optional dependency support for `kubik_wysiwyg`
+- Added `KubikMediaLibrary.wysiwyg_available?` method to check if WYSIWYG functionality is available
+- Updated media upload forms to conditionally use WYSIWYG editors when available
+
 ## [0.1.4] - 2021-06-04
 
 - Changes to syntax for instantiation methods in model